refactor: move line helper out of file.rs
